$SCRIPT_ROOT var for api urls
Add a $SCRIPT_ROOT variable in front of the api urls for any ajax calls. This affords flexibility in where the api route is hosted (rather than only at default root route) while still maintaining the advantage of static caching. The other change to make this work was made in the base.html template.